+++ test/sequel/sqlite_sequence_test.rb
@@ -366,4 +366,93 @@ class SqliteSequenceTest < Minitest::Test
     SQLiteDB.setval(:position, 100)
     assert_equal 100, SQLiteDB.currval(:position)
   end
+
+  test 'deletes history from the sequence_table' do
+    with_migration do
+      def up
+        create_sequence :position_id, if_exists: false, start: 1
+      end
+    end.up
+
+    assert_equal 1, SQLiteDB.currval(:position_id)
+
+    sequence_table_size = SQLiteDB.fetch('SELECT count(*) as len FROM position_id;').first[:len]
+    assert_equal 1, sequence_table_size
+
+    10.times { SQLiteDB.nextval(:position_id) }
+
+    assert_equal 11, SQLiteDB.currval(:position_id)
+
+    sequence_table_size = SQLiteDB.fetch('SELECT count(*) as len FROM position_id;').first[:len]
+    assert_equal 11, sequence_table_size
+
+    SQLiteDB.delete_to_currval(:position_id)
+
+    assert_equal 11, SQLiteDB.currval(:position_id)
+
+    sequence_table_size = SQLiteDB.fetch('SELECT count(*) as len FROM position_id;').first[:len]
+    assert_equal 1, sequence_table_size
+  end
+
+  test 'deletes sequences by "drop_sequence?"' do
+    with_migration do
+      def up
+        create_sequence :a
+        create_sequence :b
+        create_sequence :c
+      end
+    end.up
+
+    assert SQLiteDB.custom_sequence?(:a)
+    assert SQLiteDB.custom_sequence?(:b)
+    assert SQLiteDB.custom_sequence?(:c)
+
+    SQLiteDB.drop_sequence?(:a, :b, :c, :d)
+
+    assert !SQLiteDB.custom_sequence?(:a)
+    assert !SQLiteDB.custom_sequence?(:b)
+    assert !SQLiteDB.custom_sequence?(:c)
+  end
+
+  test 'deletes if exists before creating! the sequence' do
+    with_migration do
+      def up
+        create_sequence! :position, { start: 1000, increment: 10 }
+      end
+    end.up
+
+    assert_equal 1000, SQLiteDB.currval(:position)
+
+    10.times { SQLiteDB.nextval(:position) }
+    assert_equal 1010, SQLiteDB.currval(:position)
+
+    SQLiteDB.create_sequence!(:position, { start: 1, increment: 1 })
+
+    assert_equal 1, SQLiteDB.currval(:position)
+
+    10.times { SQLiteDB.nextval(:position) }
+    assert_equal 11, SQLiteDB.currval(:position)
+  end
+
+  test 'checks the BIGINT primery key for a sequence table' do
+    with_migration do
+      def up
+        create_sequence :position, { start: 9_223_372_036_854_775_800 }
+      end
+    end.up
+
+    assert_equal 9_223_372_036_854_775_800, SQLiteDB.currval(:position)
+    assert_equal 9_223_372_036_854_775_801, SQLiteDB.nextval(:position)
+  end
+
+  test 'checks the BIGINT primery key with negative value for a sequence table (does not support)' do
+    with_migration do
+      def up
+        create_sequence :position, { start: -9_223_372_036_854_775_807 }
+      end
+    end.up
+
+    assert_equal 0, SQLiteDB.currval(:position)
+    assert_equal 1, SQLiteDB.nextval(:position)
+  end
 end